Development

+++ mac/scripts/mac-catalina.sh
@@ -1,4 +1,51 @@
 #!/bin/zsh
+#                    _           _        _ _ 
+#  ___  _____  __   (_)_ __  ___| |_ __ _| | |
+# / _ \/ __\ \/ /   | | '_ \/ __| __/ _` | | |
+#| (_) \__ \>  <    | | | | \__ \ || (_| | | |
+# \___/|___/_/\_\   |_|_| |_|___/\__\__,_|_|_|
+
+
+# Set the colours you can use
+black=$(tput setaf 0)
+red=$(tput setaf 1)
+green=$(tput setaf 2)
+yellow=$(tput setaf 3)
+blue=$(tput setaf 4)
+magenta=$(tput setaf 5)
+cyan=$(tput setaf 6)
+white=$(tput setaf 7)
+echo "Mac OS X install script" 
+sleep 3 
+echo "By luke-h1" 
+sleep 3 
+echo "follow me on github... https://github.com/luke-h1" 
+sleep 3 
+echo ""
+echo "###############################################" $red
+echo "#        DO NOT RUN THIS SCRIPT BLINDLY       #" $red
+echo "#         YOU'LL PROBABLY REGRET IT...        #" $red
+echo "#                                             #" $red
+echo "#              READ IT THOROUGHLY             #" $red
+echo "#         AND EDIT TO SUIT YOUR NEEDS         #" $red
+echo "###############################################" $red
+echo ""
+sleep 3 
+CONTINUE=false
+echo ""
+echo "Have you read through the script you're about to run and " $red
+echo "understood that it will make changes to your computer? (y/n) ? " $red
+read -r response
+if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]; then
+  CONTINUE=true
+fi
+
+if ! $CONTINUE; then
+  # Check if we're continuing and output a message if not
+  cecho "Please go read the script, it only takes a few minutes" $red
+  exit
+fi
+
 xcode-select --install
 /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
 ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)" < /dev/null 2> /dev/null ; brew install caskroom/cask/brew-cask 2> /dev/null
@@ -15,7 +62,7 @@ cd /Users/lukehowsam/.nvm
 sudo git fetch
 brew services start openvpn
 brew services start mongodb-community@4.2  
-zsh --login 
+bash --login 
 echo "Enter your github username:"
 read user 
 git config --global --replace-all user.name "$user" 
@@ -23,6 +70,204 @@ echo "Enter your github email:"
 read email 
 git config --global --replace-all user.email "$email" 
 echo "Your github information has now been configured globally.."
-echo "script finished 😀 "  
-sleep 3 
-exit 55 
+
+
+
+# Save to disk (not to iCloud) by default
+defaults write NSGlobalDomain NSDocumentSaveNewDocumentsToCloud -bool false
+defaults write com.apple.screensaver askForPassword -int 1
+defaults write com.apple.screensaver askForPasswordDelay -int 0
+
+# Disable shadow in screenshots
+defaults write com.apple.screencapture disable-shadow -bool true
+
+###############################################################################
+# Mac App Store                                                               #
+###############################################################################
+
+# Enable the automatic update check
+defaults write com.apple.SoftwareUpdate AutomaticCheckEnabled -bool true
+
+# Download newly available updates in background
+defaults write com.apple.SoftwareUpdate AutomaticDownload -int 1
+
+# Install System data files & security updates
+defaults write com.apple.SoftwareUpdate CriticalUpdateInstall -int 1
+
+
+echo ""
+echo "Would you like to set your computer name (as done via System Preferences >> Sharing)?  (y/n)"
+read -r response
+if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]; then
+  echo "What would you like it to be?"
+  read COMPUTER_NAME
+  sudo scutil --set ComputerName $COMPUTER_NAME
+  sudo scutil --set HostName $COMPUTER_NAME
+  sudo scutil --set LocalHostName $COMPUTER_NAME
+  sudo defaults write /Library/Preferences/SystemConfiguration/com.apple.smb.server NetBIOSName -string $COMPUTER_NAME
+fi
+
+echo ""
+echo "Turn off keyboard illumination when computer is not used for 5 minutes"
+defaults write com.apple.BezelServices kDimTime -int 300
+
+echo ""
+echo "Disable display from automatically adjusting brightness? (y/n)"
+read -r response
+if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]; then
+  sudo defaults write /Library/Preferences/com.apple.iokit.AmbientLightSensor "Automatic Display Enabled" -bool false
+fi
+
+echo ""
+echo "Where do you want screenshots to be stored? (hit ENTER if you want ~/Desktop as default)"
+# Thanks https://github.com/omgmog
+read screenshot_location
+echo ""
+
+f [ -z "${screenshot_location}" ]
+then
+  # If nothing specified, we default to ~/Desktop
+  screenshot_location="${HOME}/Desktop"
+else
+  # Otherwise we use input
+  if [[ "${screenshot_location:0:1}" != "/" ]]
+  then
+    # If input doesn't start with /, assume it's relative to home
+    screenshot_location="${HOME}/${screenshot_location}"
+  fi
+fi
+echo "Setting location to ${screenshot_location}"
+defaults write com.apple.screencapture location -string "${screenshot_location}"
+
+
+echo ""
+echo "What format should screenshots be saved as? (hit ENTER for PNG, options: BMP, GIF, JPG, PDF, TIFF) "
+read screenshot_format
+if [ -z "$1" ]
+then
+  echo ""
+  echo "Setting screenshot format to PNG"
+  defaults write com.apple.screencapture type -string "png"
+else
+  echo ""
+  echo "Setting screenshot format to $screenshot_format"
+  defaults write com.apple.screencapture type -string "$screenshot_format"
+fi
+
+
+echo ""
+echo "Show hidden files in Finder by default? (y/n)"
+read -r response
+if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]; then
+  defaults write com.apple.Finder AppleShowAllFiles -bool true
+fi
+
+
+###############################################################################
+# Chrome, Safari, & WebKit
+###############################################################################
+
+echo ""
+echo "Privacy: Don't send search queries to Apple"
+defaults write com.apple.Safari UniversalSearchEnabled -bool false
+defaults write com.apple.Safari SuppressSearchSuggestions -bool true
+
+echo ""
+echo "Hiding Safari's bookmarks bar by default"
+defaults write com.apple.Safari ShowFavoritesBar -bool false
+
+echo ""
+echo "Hiding Safari's sidebar in Top Sites"
+defaults write com.apple.Safari ShowSidebarInTopSites -bool false
+
+echo ""
+echo "Disabling Safari's thumbnail cache for History and Top Sites"
+defaults write com.apple.Safari DebugSnapshotsUpdatePolicy -int 2
+
+echo ""
+echo "Enabling Safari's debug menu"
+defaults write com.apple.Safari IncludeInternalDebugMenu -bool true
+
+echo ""
+echo "Making Safari's search banners default to Contains instead of Starts With"
+defaults write com.apple.Safari FindOnPageMatchesWordStartsOnly -bool false
+
+echo ""
+echo "Removing useless icons from Safari's bookmarks bar"
+defaults write com.apple.Safari ProxiesInBookmarksBar "()"
+
+
+#########################################################################
+# Transmission.app                                                            #
+###############################################################################
+
+
+echo ""
+echo "Do you use Transmission for torrenting? (y/n)"
+read -r response
+if [[ $response =~ ^([yY][eE][sS]|[yY])$ ]]; then
+  mkdir -p ~/Downloads/Incomplete
+
+  echo ""
+  echo "Setting up an incomplete downloads folder in Downloads"
+  defaults write org.m0k.transmission UseIncompleteDownloadFolder -bool true
+  defaults write org.m0k.transmission IncompleteDownloadFolder -string "${HOME}/Downloads/Incomplete"
+
+  echo ""
+  echo "Setting auto-add folder to be Downloads"
+  defaults write org.m0k.transmission AutoImportDirectory -string "${HOME}/Downloads"
+
+  echo ""
+  echo "Don't prompt for confirmation before downloading"
+  defaults write org.m0k.transmission DownloadAsk -bool false
+
+  echo ""
+  echo "Trash original torrent files after adding them"
+  defaults write org.m0k.transmission DeleteOriginalTorrent -bool true
+
+  echo ""
+  echo "Hiding the donate message"
+  defaults write org.m0k.transmission WarningDonate -bool false
+
+  echo ""
+  echo "Hiding the legal disclaimer"
+  defaults write org.m0k.transmission WarningLegal -bool false
+
+  echo ""
+  echo "Auto-resizing the window to fit transfers"
+  defaults write org.m0k.transmission AutoSize -bool true
+
+  echo ""
+  echo "Auto updating to betas"
+  defaults write org.m0k.transmission AutoUpdateBeta -bool true
+
+  echo ""
+  echo "Setting up the best block list"
+  defaults write org.m0k.transmission EncryptionRequire -bool true
+  defaults write org.m0k.transmission BlocklistAutoUpdate -bool true
+  defaults write org.m0k.transmission BlocklistNew -bool true
+  defaults write org.m0k.transmission BlocklistURL -string "http://john.bitsurge.net/public/biglist.p2p.gz"
+fi
+
+
+###############################################################################
+# Kill affected applications
+###############################################################################
+
+echo ""
+cecho "Done!" $cyan
+echo ""
+echo ""
+cecho "################################################################################" $white
+echo ""
+echo ""
+cecho "Note that some of these changes require a logout/restart to take effect." $red
+cecho "Killing some open applications in order to take effect." $red
+echo ""
+
+find ~/Library/Application\ Support/Dock -name "*.db" -maxdepth 1 -delete
+for app in "Activity Monitor" "Address Book" "Calendar" "Contacts" "cfprefsd" \
+  "Dock" "Finder" "Mail" "Messages" "Safari" "SystemUIServer" \
+  "Terminal" "Transmission"; do
+  killall "${app}" > /dev/null 2>&1
+done
